+++ sdk/python/v1alpha1/test/test_container_volume.py
@@ -19,6 +19,7 @@ import fluxoperator
 from fluxoperator.models.container_volume import ContainerVolume  # noqa: E501
 from fluxoperator.rest import ApiException
 
+
 class TestContainerVolume(unittest.TestCase):
     """ContainerVolume unit test stubs"""
 
@@ -30,24 +31,22 @@ class TestContainerVolume(unittest.TestCase):
 
     def make_instance(self, include_optional):
         """Test ContainerVolume
-            include_option is a boolean, when False only required
-            params are included, when True both required and
-            optional params are included """
+        include_option is a boolean, when False only required
+        params are included, when True both required and
+        optional params are included"""
         # model = fluxoperator.models.container_volume.ContainerVolume()  # noqa: E501
-        if include_optional :
+        if include_optional:
+            return ContainerVolume(path="", read_only=True)
+        else:
             return ContainerVolume(
-                path = '', 
-                read_only = True
+                path="",
             )
-        else :
-            return ContainerVolume(
-                path = '',
-        )
 
     def testContainerVolume(self):
         """Test ContainerVolume"""
         inst_req_only = self.make_instance(include_optional=False)
         inst_req_and_optional = self.make_instance(include_optional=True)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()