+++ sdk/python/v1alpha1/test/test_mini_cluster_volume.py
@@ -19,6 +19,7 @@ import fluxoperator
 from fluxoperator.models.mini_cluster_volume import MiniClusterVolume  # noqa: E501
 from fluxoperator.rest import ApiException
 
+
 class TestMiniClusterVolume(unittest.TestCase):
     """MiniClusterVolume unit test stubs"""
 
@@ -30,33 +31,30 @@ class TestMiniClusterVolume(unittest.TestCase):
 
     def make_instance(self, include_optional):
         """Test MiniClusterVolume
-            include_option is a boolean, when False only required
-            params are included, when True both required and
-            optional params are included """
+        include_option is a boolean, when False only required
+        params are included, when True both required and
+        optional params are included"""
         # model = fluxoperator.models.mini_cluster_volume.MiniClusterVolume()  # noqa: E501
-        if include_optional :
+        if include_optional:
             return MiniClusterVolume(
-                annotations = {
-                    'key' : ''
-                    }, 
-                capacity = '', 
-                labels = {
-                    'key' : ''
-                    }, 
-                path = '', 
-                secret = '', 
-                secret_namespace = '', 
-                storage_class = ''
+                annotations={"key": ""},
+                capacity="",
+                labels={"key": ""},
+                path="",
+                secret="",
+                secret_namespace="",
+                storage_class="",
             )
-        else :
+        else:
             return MiniClusterVolume(
-                path = '',
-        )
+                path="",
+            )
 
     def testMiniClusterVolume(self):
         """Test MiniClusterVolume"""
         inst_req_only = self.make_instance(include_optional=False)
         inst_req_and_optional = self.make_instance(include_optional=True)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()