+++ sdk/python/v1alpha1/test/test_pod_spec.py
@@ -19,6 +19,7 @@ import fluxoperator
 from fluxoperator.models.pod_spec import PodSpec  # noqa: E501
 from fluxoperator.rest import ApiException
 
+
 class TestPodSpec(unittest.TestCase):
     """PodSpec unit test stubs"""
 
@@ -30,30 +31,22 @@ class TestPodSpec(unittest.TestCase):
 
     def make_instance(self, include_optional):
         """Test PodSpec
-            include_option is a boolean, when False only required
-            params are included, when True both required and
-            optional params are included """
+        include_option is a boolean, when False only required
+        params are included, when True both required and
+        optional params are included"""
         # model = fluxoperator.models.pod_spec.PodSpec()  # noqa: E501
-        if include_optional :
+        if include_optional:
             return PodSpec(
-                annotations = {
-                    'key' : ''
-                    }, 
-                labels = {
-                    'key' : ''
-                    }, 
-                resources = {
-                    'key' : None
-                    }
+                annotations={"key": ""}, labels={"key": ""}, resources={"key": None}
             )
-        else :
-            return PodSpec(
-        )
+        else:
+            return PodSpec()
 
     def testPodSpec(self):
         """Test PodSpec"""
         inst_req_only = self.make_instance(include_optional=False)
         inst_req_and_optional = self.make_instance(include_optional=True)
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()